% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar_dois_numeros.R
\name{somar_dois_numeros}
\alias{somar_dois_numeros}
\title{Somar dois numeros}
\usage{
somar_dois_numeros(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Outro numero}
}
\value{
Um numero
}
\description{
Somar dois numeros
}
\examples{
somar_dois_numeros(1, 1)
}
